Support local terminals on Windows, use Ruff formatter instead of Black
